Move benches to their own file, use bencher so rust stable works
