+++ Cargo.toml
@@ -11,3 +11,10 @@ repository = "https://github.com/erickt/rust-mdbm"
 mdbm-sys = { version = "*", path = "mdbm-sys" }
 libc = "0.2"
 
+[dev-dependencies]
+"bencher" = "0.1.5"
+
+[[bench]]
+name = "benches"
+path = "src/benches.rs"
+harness = false
